+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 * `measure.time_remaining` will not return negative values
 * `MemoryErrorExceptionHandler` will use less memory procesing `MemoryError`s
+* `merge_series` will now throw a `ValueError` instead of `RuntimeError`